feat: code cleanup
