+++ packages/indexer/src/api/endpoints/activities/get-user-activity/v6.ts
@@ -4,7 +4,7 @@ import _ from "lodash";
 import { Request, RouteOptions } from "@hapi/hapi";
 import Joi from "joi";
 import { logger } from "@/common/logger";
-import { fromBuffer, regex } from "@/common/utils";
+import { regex } from "@/common/utils";
 import { CollectionSets } from "@/models/collection-sets";
 import * as Boom from "@hapi/boom";
 import {
@@ -24,12 +24,11 @@ import { ActivityType } from "@/elasticsearch/indexes/activities/base";
 import * as ActivitiesIndex from "@/elasticsearch/indexes/activities";
 import * as Sdk from "@reservoir0x/sdk";
 import { Collections } from "@/models/collections";
-import { redis } from "@/common/redis";
-import { redb } from "@/common/db";
 import { Sources } from "@/models/sources";
 import { MetadataStatus } from "@/models/metadata-status";
 import { Assets } from "@/utils/assets";
 import { ActivitiesCollectionCache } from "@/models/activities-collection-cache";
+import { ActivitiesTokenCache } from "@/models/activities-token-cache";
 
 const version = "v6";
 
@@ -243,104 +242,9 @@ export const getUserActivityV6Options: RouteOptions = {
         );
 
         try {
-          let tokensToFetch = activities
-            .filter((activity) => activity.token)
-            .map((activity) => `token-cache:${activity.contract}:${activity.token?.id}`);
-
-          if (tokensToFetch.length) {
-            // Make sure each token is unique
-            tokensToFetch = [...new Set(tokensToFetch).keys()];
-
-            tokensMetadata = await redis.mget(tokensToFetch);
-            tokensMetadata = tokensMetadata
-              .filter((token) => token)
-              .map((token) => JSON.parse(token));
-
-            const nonCachedTokensToFetch = tokensToFetch.filter((tokenToFetch) => {
-              const [, contract, tokenId] = tokenToFetch.split(":");
-
-              return (
-                tokensMetadata.find((token) => {
-                  return token.contract === contract && token.token_id === tokenId;
-                }) === undefined
-              );
-            });
-
-            if (nonCachedTokensToFetch.length) {
-              const tokensFilter = [];
-
-              for (const nonCachedTokenToFetch of nonCachedTokensToFetch) {
-                const [, contract, tokenId] = nonCachedTokenToFetch.split(":");
-
-                tokensFilter.push(`('${_.replace(contract, "0x", "\\x")}', '${tokenId}')`);
-              }
-
-              // Fetch details for all tokens
-              const tokensResult = await redb.manyOrNone(
-                `
-          SELECT
-            tokens.contract,
-            tokens.token_id,
-            tokens.name,
-            tokens.image,
-            tokens.metadata_disabled,
-            tokens.image_version,
-            (tokens.metadata->>'image_mime_type') as image_mime_type,
-            tokens.rarity_score,
-            tokens.rarity_rank
-          FROM tokens
-          WHERE (tokens.contract, tokens.token_id) IN ($/tokensFilter:raw/)
-        `,
-                { tokensFilter: _.join(tokensFilter, ",") }
-              );
-
-              if (tokensResult?.length) {
-                tokensMetadata = tokensMetadata.concat(
-                  tokensResult.map((token) => ({
-                    contract: fromBuffer(token.contract),
-                    token_id: token.token_id,
-                    name: token.name,
-                    image: token.image,
-                    image_version: token.image_version,
-                    image_mime_type: token.image_mime_type,
-                    metadata_disabled: token.metadata_disabled,
-                    rarity_score: token.rarity_score,
-                    rarity_rank: token.rarity_rank,
-                  }))
-                );
-
-                const redisMulti = redis.multi();
-
-                for (const tokenResult of tokensResult) {
-                  const tokenResultContract = fromBuffer(tokenResult.contract);
-
-                  await redisMulti.set(
-                    `token-cache:${tokenResultContract}:${tokenResult.token_id}`,
-                    JSON.stringify({
-                      contract: tokenResultContract,
-                      token_id: tokenResult.token_id,
-                      name: tokenResult.name,
-                      image: tokenResult.image,
-                      image_version: tokenResult.image_version,
-                      image_mime_type: tokenResult.image_mime_type,
-                      metadata_disabled: tokenResult.metadata_disabled,
-                      rarity_score: tokenResult.rarity_score,
-                      rarity_rank: tokenResult.rarity_rank,
-                    })
-                  );
-
-                  await redisMulti.expire(
-                    `token-cache:${tokenResultContract}:${tokenResult.token_id}`,
-                    60 * 60 * 24
-                  );
-                }
-
-                await redisMulti.exec();
-              }
-            }
-          }
+          tokensMetadata = await ActivitiesTokenCache.getTokens(activities);
         } catch (error) {
-          logger.error(`get-user-activity-${version}-handler`, `Token cache error: ${error}`);
+          logger.error(`get-collection-activity-${version}-handler`, `Token cache error: ${error}`);
         }
 
         try {
