+++ packages/indexer/src/elasticsearch/indexes/asks/event-handlers/base/index.ts
@@ -25,15 +25,12 @@ export abstract class BaseAskEventHandler {
       contract: data.contract,
       token_id: data.token_id,
       token_name: data.token_name,
-      token_image: data.token_image,
-      token_media: data.token_media,
       token_is_flagged: Number(data.token_is_flagged),
       token_is_spam: Number(data.token_is_spam),
       token_rarity_rank: data.token_rarity_rank ? Number(data.token_rarity_rank) : undefined,
       token_attributes: data.token_attributes,
       collection_id: data.collection_id,
       collection_name: data.collection_name,
-      collection_image: data.collection_image,
       collection_is_spam: Number(data.collection_is_spam),
       order_id: data.order_id,
       order_source_id_int: Number(data.order_source_id_int),
